+++ packages/database/src/seed/index.ts
@@ -1,3 +1,3 @@
-export * from "./demo";
 export * from "./modules";
+export * from "./seed-demo-data";
 export * from "./stripe";
